requirements.txt
kafka-python==2.0.2
pandas==2.2.2
pyarrow==16.1.0
PyYAML==6.0.2
python-dateutil==2.9.0.post0
pyspark==3.5.2
delta-spark==3.2.0
pytest==8.3.2
black==24.8.0
ruff==0.6.3
